FixtureTestCaseTrait
